+++ Lib/fontbakery/reporters/ghmarkdown.py
@@ -10,7 +10,7 @@ LOGLEVELS=["ERROR","FAIL","WARN","SKIP","INFO","PASS","DEBUG"]
 class GHMarkdownReporter(SerializeReporter):
 
     def write(self):
-        with open(self.output_file, "w") as fh:
+        with open(self.output_file, "w", encoding="utf8") as fh:
             fh.write(self.get_markdown())
         print(f'A report in GitHub Markdown format which can be useful\n'
               f' for posting issues on a GitHub issue tracker has been\n'