+++ CHANGELOG.md
@@ -2,7 +2,9 @@ Below are the most important changes from each release.
 A more detailed list of changes is available in the corresponding milestones for each release in the Github issue tracker (https://github.com/googlefonts/fontbakery/milestones?state=closed).
 
 ## Upcoming release: 0.8.12 (2023-Apr-??)
-  - ...
+### New Checks
+#### Added to the Open Type Profile
+  - **[com.google.fonts/check/name/italic_names]:** Implemented checks for name IDs 1, 2, 16, 17 for Italic fonts (issues #3666 and #3667)
 
 
 ## 0.8.11 (2023-Mar-03)