+++ CHANGELOG.md
@@ -6,6 +6,8 @@ A more detailed list of changes is available in the corresponding milestones for
 #### Added to the Open Type Profile
   - **[com.google.fonts/check/name/italic_names]:** Implemented checks for name IDs 1, 2, 16, 17 for Italic fonts (issues #3666 and #3667)
 
+### Tests
+  - Added test for com.google.fonts/soft_doted. (issue #4069)
 
 ## 0.8.11 (2023-Mar-03)
 ### Noteworthy code-changes